+++ inst/shinybrms_app/app.R
@@ -1,5 +1,5 @@
 # shinybrms: Graphical User Interface ('shiny' App) for 'brms'
-# Copyright (C) 2020  Frank Weber
+# Copyright (C) 2021  Frank Weber
 #   
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -335,18 +335,23 @@ ui <- navbarPage(
         strong("Parameters (with corresponding link functions) specific to this distributional family:"),
         tableOutput("dist_link"),
         helpText(
-          "For details concerning the link functions, see the help for the R function",
-          a(HTML(paste(code("brms::brmsfamily()"))),
-            href = "https://paul-buerkner.github.io/brms/reference/brmsfamily.html",
-            target = "_blank"),
-          "and the", strong("brms"), "vignette",
-          a("\"Parameterization of Response Distributions in brms\"",
-            href = "https://CRAN.R-project.org/web/packages/brms/vignettes/brms_families.html",
-            target = "_blank",
-            .noWS = "after"),
-          ". Note that for each parameter, the link function only applies if this parameter is",
-          "actually modeled by (nonconstant) predictors. Currently, this is only supported",
-          "for the location parameter (e.g.", code("mu"), "for a Gaussian distribution)."
+          p("For details concerning the link functions, see the help for the R function",
+            a(HTML(paste(code("brms::brmsfamily()"))),
+              href = "https://paul-buerkner.github.io/brms/reference/brmsfamily.html",
+              target = "_blank"),
+            "and the", strong("brms"), "vignette",
+            a("\"Parameterization of Response Distributions in brms\"",
+              href = "https://CRAN.R-project.org/web/packages/brms/vignettes/brms_families.html",
+              target = "_blank",
+              .noWS = "after"),
+            ". Note that for each parameter, the link function only applies if this parameter is",
+            "actually modeled by (nonconstant) predictors. Currently, this is only supported",
+            "for the location parameter (e.g.", code("mu"), "for a Gaussian distribution)."),
+          p("For details concerning the remaining (family-specific) parameters, see the help for the R function ",
+            a(HTML(paste(code("brms::set_prior()"))),
+              href = "https://paul-buerkner.github.io/brms/reference/set_prior.html",
+              target = "_blank"),
+            ".")
         )
       ),
       tabPanel(
@@ -506,9 +511,20 @@ ui <- navbarPage(
         tags$li("When specifying a custom prior, you may only choose a combination of",
                 "\"Class\", \"Coefficient\", and \"Group\" which is also present in the",
                 "table of the default priors."),
-        tags$li("The parameter named", code("Intercept"), "is the intercept when centering the predictors.",
-                "This is only the internally used intercept; in the output, the intercept with",
-                "respect to the noncentered predictors is given (named", code("b_Intercept", .noWS = "after"), ")."),
+        tags$li("The names of the parameter classes are taken from", strong("brms"),
+                "and may be translated as follows:",
+                tags$ul(
+                  tags$li(code("Intercept"), ": the intercept when centering the predictors.",
+                          "This is only the internally used intercept; in the output, the intercept with",
+                          "respect to the noncentered predictors is given (named", code("b_Intercept", .noWS = "after"), ")."),
+                  tags$li(code("b"), ": nonpooled effects (or nonpooled regression coefficients)."),
+                  tags$li(code("sd"), ": standard deviations of partially pooled effects."),
+                  tags$li(code("cor"), ": correlations between partially pooled effects of the same group."),
+                  tags$li("All other parameter classes are specific to the chosen",
+                          "distributional family for the outcome (see page",
+                          HTML(paste(actionLink("outcome_link1", HTML("Likelihood &rarr; Outcome"))), .noWS = "after"),
+                          ").")
+                )),
         tags$li("As soon as you choose a new dataset on page", actionLink("data_link3", "Data"), "(even if you upload",
                 "the same dataset again), the custom priors are automatically reset."),
         tags$li("As soon as you change the likelihood, the custom priors are automatically reset.")
@@ -665,7 +681,7 @@ ui <- navbarPage(
           checkboxInput("show_stancode", "Show Stan code", value = FALSE),
           conditionalPanel(
             condition = "input.show_stancode",
-            verbatimTextOutput("stancode_view", placeholder = TRUE),
+            verbatimTextOutput("stancode_view", placeholder = TRUE)
           ),
           downloadButton("stancode_download", "Download Stan code")
         ),
@@ -737,7 +753,8 @@ ui <- navbarPage(
                      numericInput("advOpts_seed", "Seed:",
                                   value = NA, step = 1L),
                      numericInput("advOpts_cores", "Cores:",
-                                  value = getOption("mc.cores", parallel::detectCores()), step = 1L, min = 1L),
+                                  value = getOption("mc.cores", parallel::detectCores(logical = FALSE)),
+                                  step = 1L, min = 1L),
                      numericInput("advOpts_chains", "Chains:",
                                   value = 4L, step = 1L, min = 1L),
                      numericInput("advOpts_iter", "Total iterations per chain:",
@@ -827,19 +844,19 @@ ui <- navbarPage(
               target = "_blank",
               .noWS = "after"),
             ",",
-            a("Betancourt (2017)",
-              href = "https://arxiv.org/abs/1701.02434",
+            a("Betancourt (2018)",
+              href = "https://arxiv.org/abs/1701.02434v2",
               target = "_blank",
               .noWS = "after"),
             ", and",
-            a("Vehtari et al. (2020)",
+            a("Vehtari et al. (2021)",
               href = "https://doi.org/10.1214/20-BA1221",
               target = "_blank",
               .noWS = "after"),
             ". The \"Brief Guide to Stan’s Warnings\" covers all MCMC diagnostics used here and",
             "gives some advice on what to do when they indicate problems.",
-            "Betancourt (2017) focuses on the HMC-specific diagnostics,",
-            "whereas Vehtari et al. (2020) focus on the general MCMC diagnostics."),
+            "Betancourt (2018) focuses on the HMC-specific diagnostics,",
+            "whereas Vehtari et al. (2021) focus on the general MCMC diagnostics."),
           p("The HMC-specific diagnostics are:",
             tags$ul(
               tags$li("the number of iterations ending with a divergence,"),
@@ -868,10 +885,7 @@ ui <- navbarPage(
                 #   target = "_blank",
                 #   .noWS = "after"),
                 # ")",
-                "proposed by",
-                a("Vehtari et al. (2020)",
-                  href = "https://doi.org/10.1214/20-BA1221",
-                  target = "_blank"),
+                "proposed by Vehtari et al. (2021)",
                 "(here simply called", em("the"), "\\(\\widehat{R}\\) instead of",
                 em("the modified"), "\\(\\widehat{R}\\)),"
               ),
@@ -919,8 +933,8 @@ ui <- navbarPage(
         )
       ),
       tabPanel(
-        "Summary",
-        titlePanel("Summary"),
+        "Default summary",
+        titlePanel("Default summary"),
         br(),
         helpText("Notes:",
                  tags$ul(
@@ -1040,7 +1054,9 @@ ui <- navbarPage(
             tags$ul(
               tags$li(
                 "In the", strong("shinystan"), "app, the parameter names given by", strong("brms"),
-                "are used. These are as follows:",
+                "are used. These are closely related to the parameter classes listed on page",
+                HTML(paste(actionLink("prior_link3", "Prior")), .noWS = "after"), "and may be",
+                "summarized as follows:",
                 tags$ul(
                   tags$li(code("b_Intercept"), "is the intercept (with respect to the noncentered predictors)."),
                   tags$li("The parameters starting with", code("b_"), "are the nonpooled effects."),
@@ -1055,7 +1071,7 @@ ui <- navbarPage(
                   tags$li(code("log-posterior"), "is the accumulated log-posterior density (up to an additive constant)."),
                   tags$li("All other parameters are parameters specific to the chosen",
                           "distributional family for the outcome (see page",
-                          HTML(paste(actionLink("outcome_link1", HTML("Likelihood &rarr; Outcome"))), .noWS = "after"),
+                          HTML(paste(actionLink("outcome_link2", HTML("Likelihood &rarr; Outcome"))), .noWS = "after"),
                           ").")
                 )
               ),
@@ -1108,14 +1124,14 @@ ui <- navbarPage(
                   "Twitter, Inc. (for Bootstrap, the basis for the Bootswatch theme \"United\");",
                   "Google, LLC (for the \"Open Sans\" font)"),
           tags$li(strong("Version:"),
-                  "1.5.0"),
+                  "1.5.1"),
           tags$li(strong("Date:"),
-                  "November 20, 2020"),
+                  "March 6, 2021"),
           tags$li(strong("Citation:"),
-                  "Frank Weber (2020).",
+                  "Frank Weber (2021).",
                   em("shinybrms: Graphical User Interface ('shiny' App)",
                      "for 'brms'."),
-                  "R package, version 1.5.0. URL:",
+                  "R package, version 1.5.1. URL:",
                   a("https://fweber144.github.io/shinybrms/",
                     href = "https://fweber144.github.io/shinybrms/",
                     target = "_blank",
@@ -1235,11 +1251,11 @@ ui <- navbarPage(
       "References",
       titlePanel("References"),
       br(),
-      p("Betancourt M (2017).",
+      p("Betancourt M (2018).",
         "A conceptual introduction to Hamiltonian Monte Carlo.",
-        em("arXiv:170102434 [stat]", .noWS = "after"), ". URL:",
-        a("https://arxiv.org/abs/1701.02434",
-          href = "https://arxiv.org/abs/1701.02434",
+        em("arXiv:170102434v2 [stat]", .noWS = "after"), ". URL:",
+        a("https://arxiv.org/abs/1701.02434v2",
+          href = "https://arxiv.org/abs/1701.02434v2",
           target = "_blank"),
         "(visited on July 26, 2020)."),
       p("Gelman A, Carlin JB, Stern HS, Dunson DB, Vehtari A, and Rubin DB (2014).",
@@ -1249,9 +1265,10 @@ ui <- navbarPage(
         em("Statistical Rethinking: A Bayesian Course with Examples in R and Stan", .noWS = "after"),
         ". 2nd ed. Boca Raton, FL: CRC Press."),
       p(withMathJax(
-        "Vehtari A, Gelman A, Simpson D, Carpenter B, and B", HTML("&uuml;", .noWS = "outside"), "rkner P-C (2020).",
+        "Vehtari A, Gelman A, Simpson D, Carpenter B, and B", HTML("&uuml;", .noWS = "outside"), "rkner P-C (2021).",
         "Rank-normalization, folding, and localization: An improved \\(\\widehat{R}\\) for",
-        "assessing convergence of MCMC.", em("Bayesian Analysis", .noWS = "after"), ". DOI:",
+        "assessing convergence of MCMC.", em("Bayesian Analysis", .noWS = "after"), ",",
+        em("-1", .noWS = "after"), "(-1), 1", HTML("&#8212;", .noWS = "outside"), "28. DOI:",
         a("10.1214/20-BA1221", href = "https://doi.org/10.1214/20-BA1221", target = "_blank",
           .noWS = "after"),
         "."))
@@ -1282,10 +1299,13 @@ server <- function(input, output, session){
     updateNavbarPage(session, "navbar_ID", "Likelihood")
   })
   
-  observeEvent(input$outcome_link1, {
+  observeEvent({
+    input$outcome_link1
+    input$outcome_link2
+  }, {
     updateNavbarPage(session, "navbar_ID", "Likelihood")
     updateNavlistPanel(session, "likelihood_navlist_ID", "Outcome")
-  })
+  }, ignoreNULL = FALSE, ignoreInit = TRUE)
   
   observeEvent(input$formula_link1, {
     updateNavbarPage(session, "navbar_ID", "Likelihood")
@@ -1295,6 +1315,7 @@ server <- function(input, output, session){
   observeEvent({
     input$prior_link1
     input$prior_link2
+    input$prior_link3
   }, {
     updateNavbarPage(session, "navbar_ID", "Prior")
   }, ignoreNULL = FALSE, ignoreInit = TRUE)
@@ -2485,7 +2506,7 @@ server <- function(input, output, session){
   })
   
   #------------------------
-  # Summary
+  # Default summary
   
   output$smmry_view <- renderPrint({
     print(C_stanres()$bfit, digits = 4, robust = TRUE, priors = TRUE, prob = 0.95, mc_se = FALSE)